# Diversity.Hill

Hill numbers are found in the Diversity.Hill submodule.

Usage

Accessing the main functionality in the package is simple:

julia> using Diversity.Hill

julia> community = [10, 20, 20, 0, 3];

julia> community /= sum(community);

julia> diversities = hillnumber(community, [0, 1, 2])
3×7 DataFrame
│ Row │ measure    │ q     │ type_level │ type_name │ partition_level │ partition_name │ diversity │
│     │ String     │ Int64 │ String     │ String    │ String          │ String         │ Float64   │
├─────┼────────────┼───────┼────────────┼───────────┼─────────────────┼────────────────┼───────────┤
│ 1   │ HillNumber │ 0     │ types      │           │ subcommunity    │ 1              │ 4.0       │
│ 2   │ HillNumber │ 1     │ types      │           │ subcommunity    │ 1              │ 3.36264   │
│ 3   │ HillNumber │ 2     │ types      │           │ subcommunity    │ 1              │ 3.09021   │

Diversity.jl

Diversity.Hill — Module

Hill numbers are found in the Diversity.Hill package.

Diversity.Hill.hillnumber — Method

hillnumber(proportions, qs)

Calculate the Hill number (or naive diversity) of order q of population(s) with given relative proportions

Arguments:

proportions: relative proportions of different individuals / species in population (vector, or matrix where columns are individual populations)

qs: single number or vector of orders of diversity measurement

Returns:

Diversity of order qs (single number or vector of diversities)

Diversity.API

Diversity.Diversity

Diversity.Ecology

Diversity.Hill

Diversity.Jost

Diversity.individualDiversity

Diversity.metacommunityDiversity

Diversity.subcommunityDiversity

Diversity.API.AbstractMetacommunity

Diversity.API.AbstractPartition

Diversity.API.AbstractTypes

Diversity.DiversityLevel

Diversity.DiversityMeasure

Diversity.Gamma

Diversity.GeneralTypes

Diversity.Metacommunity

Diversity.NormalisedAlpha

Diversity.NormalisedBeta

Diversity.NormalisedRho

Diversity.Onecommunity

Diversity.PowerMeanMeasure

Diversity.RawAlpha

Diversity.RawBeta

Diversity.RawRho

Diversity.RelativeEntropyMeasure

Diversity.Species

Diversity.Subcommunities

Diversity.Taxonomy

Diversity.UniqueTypes

Diversity.API._addedoutputcols

Diversity.API._calcabundance

Diversity.API._calcordinariness

Diversity.API._calcsimilarity

Diversity.API._countsubcommunities

Diversity.API._counttypes

Diversity.API._getabundance

Diversity.API._getaddedoutput

Diversity.API._getdiversityname

Diversity.API._getmetaabundance

Diversity.API._getmetaordinariness!

Diversity.API._getordinariness!

Diversity.API._getpartition

Diversity.API._getscale

Diversity.API._getsubcommunitynames

Diversity.API._gettypenames

Diversity.API._gettypes

Diversity.API._getweight

Diversity.API.floattypes

Diversity.API.mcmatch

Diversity.API.typematch

Diversity.Ecology.generalisedjaccard

Diversity.Ecology.generalisedrichness

Diversity.Ecology.generalisedshannon

Diversity.Ecology.generalisedsimpson

Diversity.Ecology.jaccard

Diversity.Ecology.richness

Diversity.Ecology.shannon

Diversity.Ecology.simpson

Diversity.Hill.hillnumber

Diversity.Jost.jostalpha

Diversity.Jost.jostbeta

Diversity.addedoutputcols

Diversity.calcsimilarity

Diversity.countsubcommunities

Diversity.counttypes

Diversity.diversity

Diversity.getASCIIName

Diversity.getFullName

Diversity.getName

Diversity.getabundance

Diversity.getaddedoutput

Diversity.getdiversityname

Diversity.getmetaabundance

Diversity.getmetaordinariness!

Diversity.getordinariness!

Diversity.getpartition

Diversity.getsubcommunitynames

Diversity.gettypenames

Diversity.gettypes

Diversity.getweight

Diversity.hassimilarity

Diversity.inddiv

Diversity.metadiv

Diversity.qD

Diversity.qDZ

Diversity.subdiv